# docs/open-api-docs.yaml
openapi: 3.0.3
info:
  title: The Agent's user-facing API
  description: The user-facing parts of The Agent's API service (excluding system-level endpoints, chat completion, maintenance endpoints, etc.)
  version: 5.5.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your.agent.com
    description: Your Agent API
security:
  - bearerAuth: []
paths:
  /settings/{settings_type}/{resource_id}:
    get:
      summary: Get user or chat settings
      description: Retrieve settings for a specific user or chat configuration
      operationId: getSettings
      tags: [Settings]
      parameters:
        - name: settings_type
          in: path
          required: true
          schema:
            type: string
            enum: [user, chat]
          description: Type of settings to retrieve
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID (hex) or Chat ID (hex, UUID without hyphens) depending on settings_type
      responses:
        "200":
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UserSettingsResponse"
                  - $ref: "#/components/schemas/ChatSettingsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/user/{resource_id}/tools:
    get:
      summary: Get user's external tools configuration
      description: Retrieve external tools and providers configuration status
      operationId: getTools
      tags: [Settings]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Tools configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalToolsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/user/{resource_id}/products:
    get:
      summary: Get available products
      description: Retrieve the list of purchasable products
      operationId: getProducts
      tags: [Settings]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/user/{user_id_hex}:
    patch:
      summary: Update user settings
      description: Save user settings including API keys and tool choices
      operationId: saveUserSettings
      tags: [Settings]
      parameters:
        - name: user_id_hex
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsPayload"
      responses:
        "200":
          description: User settings saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /settings/chat/{chat_id}:
    patch:
      summary: Update chat settings
      description: Save chat configuration including language and reply settings
      operationId: saveChatSettings
      tags: [Settings]
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: Chat ID in hexadecimal format (UUID without hyphens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSettingsPayload"
      responses:
        "200":
          description: Chat settings saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /user/{resource_id}/chats:
    get:
      summary: Get user-administered chats
      description: Retrieve all private and group chats where the user is an administrator
      operationId: getChats
      tags: [Chats]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Chats retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatInfo"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /user/{resource_id}/sponsorships:
    get:
      summary: Get user's sponsorships
      description: Retrieve all sponsorships created by the user
      operationId: getSponsorships
      tags: [Sponsorships]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Sponsorships retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorshipsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []

    post:
      summary: Create a new sponsorship
      description: Sponsor another user by their platform-specific handle
      operationId: sponsorUser
      tags: [Sponsorships]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorshipPayload"
      responses:
        "200":
          description: User sponsored successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorshipCreateResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /user/{resource_id}/sponsorships/{platform}/{platform_handle}:
    delete:
      summary: Remove a sponsorship
      description: Remove sponsorship for a specific user
      operationId: unsponsorUser
      tags: [Sponsorships]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [telegram, whatsapp]
          description: Platform type for the handle resolution
        - name: platform_handle
          in: path
          required: true
          schema:
            type: string
          description: Platform-specific handle of the sponsored user (e.g., Telegram username, WhatsApp phone number)
      responses:
        "200":
          description: Sponsorship removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /user/{resource_id}/sponsored:
    delete:
      summary: Remove self-sponsorship
      description: User removes their own sponsorship status
      operationId: unsponsorSelf
      tags: [Sponsorships]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: User ID in hexadecimal format (UUID without hyphens)
      responses:
        "200":
          description: Self-sponsorship removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
  /user/{resource_id}/transfers:
    post:
      summary: Transfer credits to another user
      description: Transfers credits from the authenticated user to another user identified by platform handle
      operationId: transferCredits
      tags: [CreditTransfers]
      parameters:
        - name: resource_id
          in: path
          required: true
          schema:
            type: string
          description: Sender's user ID in hexadecimal format (UUID without hyphens)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditTransferPayload"
      responses:
        "200":
          description: Credits transferred successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "429":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceErrorResponse"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from one of the auth flows (usually via the chat)
  schemas:
...
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          nullable: false
          items:
            $ref: "#/components/schemas/ConfiguredProduct"
          description: List of purchasable products
      required:
        - products
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          nullable: false
          enum: [OK]
          description: Operation status
      required:
        - status
...
  responses:
    UnauthorizedError:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                nullable: false
                description: Error reason
            required:
              - detail
    ForbiddenError:
      description: Access denied — authorization failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"
    NotFoundError:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceErrorResponse"